% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-matlab.R
\name{get_matlab}
\alias{get_matlab}
\title{Get path to matlab file}
\usage{
get_matlab(path = NULL)
}
\arguments{
\item{path}{Name of file. If \code{NULL}, the example files will be listed.}
}
\description{
This function is modified from the package \code{readr} to access
the bundled datasets in directory \code{inst/extdata} of \code{readmat}. This
function make them easy to access. This function is modified from
\code{\link[readr:readr_example]{readr_example}} of the package
\code{\link[readr]{readr}}.
}
