% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-matlab-names.R
\name{get_names}
\alias{get_names}
\title{Matlab object names}
\usage{
get_names(file)
}
\arguments{
\item{file}{A path to a file. Files ending with \code{.mat}.}
}
\value{
A vector or array with up to 3 dimensions.
}
\description{
\code{get_names()} can read Matlab files
}
