% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readmat.R
\name{read_mat}
\alias{read_mat}
\title{Read a Matlab file

\code{read_mat()} can read Matlab files}
\usage{
read_mat(file)
}
\arguments{
\item{file}{A path to a file. Files ending with \code{.mat}.}
}
\value{
A vector or array with up to 3 dimensions.
}
\description{
Read a Matlab file

\code{read_mat()} can read Matlab files
}
\examples{
# Use \code{get_matlab()} to access the examples bundled with this package

# Read Matlab file with one object
\dontrun{read_mat(get_matlab("double-vec.mat"))}

# Read Matlab file with multiple objects
\dontrun{read_mat(get_matlab("multi-object.mat"))}

}
